patch(status-state): Don't show pending when unapproved
Graphite shows a clock and a spinning wheel whenever there are any
pending statuses in github. This is very confusing, as you would expect
that you can just wait to solve the problem, but in reality you need to
reach out.

We should roll this back if having a failure marker is more confusing
than having a spinning wheel.